% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barentswatch_customApi.R
\name{barentswatch_customApi}
\alias{barentswatch_customApi}
\title{Get custom API request string}
\usage{
barentswatch_customApi(request)
}
\arguments{
\item{request}{Request string}
}
\description{
Type in your own request string
}
\examples{
barentswatch_customApi(request)
}
