% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barentswatch_auth.R
\name{get_barentswatch_token}
\alias{get_barentswatch_token}
\title{Retrieve BarentsWatch token from environment and configs for httr}
\usage{
get_barentswatch_token(shiny_return_token = NULL)
}
\arguments{
\item{shiny_return_token}{In a shiny session, this is passed instead.}
}
\value{
a httr configured option for token
For shiny the token is passed from reactive session
}
\description{
Get token if it's previously stored, else prompt user to get one.
}
\concept{authentication functions}
\keyword{internal}
